feat: add reset basket state and redirect to main page
