+++ src/App.tsx
@@ -5,19 +5,25 @@ import Main from './pages/Main';
 import Basket from './pages/Basket';
 import ErrorPage from './pages/Error';
 import { createContext, useReducer, Dispatch, useEffect } from 'react';
-import { BasketState, actionType, basketReducer, initBasket } from './hooks/basketReducer';
+import {
+  BasketState,
+  actionType,
+  basketReducer,
+  initBasket,
+  initState,
+} from './hooks/basketReducer';
 import ProductDetails from './components/Product-details/Product-details';
 
 export const BasketContext = createContext<{
   basketState: BasketState;
   dispatch: Dispatch<actionType>;
 }>({
-  basketState: [],
+  basketState: initState,
   dispatch: () => null,
 });
 
 function App() {
-  const [basketState, dispatch] = useReducer(basketReducer, [], initBasket);
+  const [basketState, dispatch] = useReducer(basketReducer, initState, initBasket);
 
   useEffect(() => {
     localStorage.setItem('basketState', JSON.stringify(basketState));
